# docs/service/catalog-service/api.yaml
openapi: 3.0.3
info:
  title: catalog-service API
  version: "0.1"
  description: |
    Quản lý Catalog — Brand, Category, AttributeTemplate, Product (SPU), Variant (SKU).
    Upstream thuần túy: không consume event, chỉ publish.

    **Auth:**
    - `/api/**` — public (không cần token)
    - `/api/admin/**` — yêu cầu role `ADMIN` (bearer token từ oauth2-service)
servers:
  - url: http://localhost:8082
    description: Local dev
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          example: null

    # ── Brand ─────────────────────────────────────────────────────────────────
...
  responses:
    400:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "VALIDATION_ERROR" }
                  message: { type: string }
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
...
    422:
      description: Business rule violation
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "INPUT_TYPE_NOT_SELECT" }
                  message: { type: string }

    # ── Product ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/products/{id}:
    get:
      tags: [Product (Public)]
      summary: Chi tiết sản phẩm (chỉ PUBLISHED)
      description: Không cần auth. Cache L1+L2 (TTL 10 min). Trả 404 nếu không PUBLISHED.
      operationId: getPublishedProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Product ULID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/ProductResponse' }
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: PRODUCT_NOT_FOUND hoặc không PUBLISHED
...
  /api/seller/products:
    post:
      tags: [Product (Seller)]
      summary: Tạo product mới (DRAFT)
      description: |
        Seller tạo product. Product bắt đầu ở trạng thái DRAFT.
        Header `X-Seller-Id` được gateway inject sau khi xác thực session.
      operationId: createProduct
      parameters:
        - name: X-Seller-Id
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        "201":
          description: Created — trả về Product ULID
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: string, description: Product ULID }
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: CATEGORY_NOT_FOUND | BRAND_NOT_FOUND
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: CATEGORY_NOT_LEAF — categoryId phải trỏ tới category leaf (L3)
          # ErrorCode: CATEGORY_INACTIVE — category đã deactivate, không cho tạo Product mới
          # ErrorCode: ATTRIBUTE_NOT_IN_CATEGORY | REQUIRED_ATTRIBUTE_MISSING
          # ErrorCode: INVALID_ATTRIBUTE_VALUE | TEMPLATE_INACTIVE | ATTRIBUTE_VALUE_FORMAT_INVALID
          # ErrorCode: VARIANT_DEFINING_REQUIRES_SELECT — isVariantDefining=true trên template không phải SELECT
    get:
      tags: [Product (Seller)]
      summary: Danh sách products của seller (paged)
      operationId: listSellerProducts
      parameters:
        - name: X-Seller-Id
          in: header
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/PagedProductSummaryResponse' }
  /api/seller/products/{id}:
    get:
      tags: [Product (Seller)]
      summary: Chi tiết product (mọi status)
      operationId: getSellerProduct
      parameters:
        - name: X-Seller-Id
          in: header
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/ProductResponse' }
        "404":
          $ref: '#/components/responses/404'
    put:
      tags: [Product (Seller)]
      summary: Cập nhật thông tin product
      description: Cập nhật name, description, warrantyInfo, attributeValues. Không đổi categoryId sau khi có variant.
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProductRequest' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: PRODUCT_BLOCKED
          # ErrorCode: ATTRIBUTE_NOT_IN_CATEGORY | REQUIRED_ATTRIBUTE_MISSING
          # ErrorCode: INVALID_ATTRIBUTE_VALUE | TEMPLATE_INACTIVE | ATTRIBUTE_VALUE_FORMAT_INVALID
          # ErrorCode: ATTRIBUTE_VALUE_STILL_USED_BY_VARIANT — gỡ 1 giá trị khỏi attributeValues mà vẫn còn
          #   Variant (bất kỳ status) dùng đúng giá trị đó trong combination; deactivate Variant đó trước.
          # ErrorCode: VARIANT_DEFINING_REQUIRES_SELECT — isVariantDefining=true trên template không phải SELECT
          # ErrorCode: VARIANT_DEFINING_LOCKED_AFTER_VARIANT — đổi isVariantDefining sau khi Product đã có Variant
    delete:
      tags: [Product (Seller)]
      summary: Xoá vĩnh viễn Product (chỉ khi còn DRAFT)
      description: |
        Hard-delete — chỉ cho phép khi `status = DRAFT` (chưa từng publish, nên chắc chắn chưa có Order
        nào tham chiếu tới bất kỳ Variant nào). Cascade xoá luôn toàn bộ Variant của Product. Không phục
        hồi được — khác hẳn `unpublish` (soft, giữ nguyên dữ liệu). Dùng để dọn các bản nháp tạo nhầm/bỏ
        dở, tránh rác tồn mãi trong danh sách Product của Seller.
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: PRODUCT_NOT_DRAFT — Product đã từng publish (status khác DRAFT), không cho xoá cứng
  /api/seller/products/{id}/publish:
    post:
      tags: [Product (Seller)]
      summary: Publish product → PUBLISHED
      description: |
        Guard: phải có ít nhất 1 Variant ACTIVE. Emit `ProductPublishedEvent`.
      operationId: publishProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: PUBLISH_REQUIRES_ACTIVE_VARIANT | PRODUCT_BLOCKED
  /api/seller/products/{id}/unpublish:
    post:
      tags: [Product (Seller)]
      summary: Unpublish product → UNPUBLISHED
      description: Emit `ProductUnpublishedEvent`.
      operationId: unpublishProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: PRODUCT_BLOCKED
  /api/seller/products/{id}/images/upload-url:
    post:
      tags: [Product (Seller)]
      summary: Lấy presigned PUT URL để upload ảnh lên MinIO
      description: |
        Presigned URL TTL = 5 phút. Browser upload thẳng lên MinIO,
        sau đó gọi /confirm để đăng ký vào Product.
      operationId: getProductImageUploadUrl
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      uploadUrl: { type: string, format: uri }
                      objectKey: { type: string }
                      ttlSeconds: { type: integer, example: 300 }
  /api/seller/products/{id}/images/confirm:
    post:
      tags: [Product (Seller)]
      summary: Confirm upload ảnh — append vào Product.images
      description: catalog-service verify object tồn tại trên MinIO trước khi lưu.
      operationId: confirmProductImageUpload
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectKey]
              properties:
                objectKey: { type: string }
      responses:
        "201":
          description: Created — trả về imageId
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: string, description: Image ULID }
        "422":
          $ref: '#/components/responses/422'
          # Object chưa tồn tại trên MinIO
  /api/seller/products/{id}/images/{imageId}:
    delete:
      tags: [Product (Seller)]
      summary: Xóa ảnh khỏi Product
      operationId: removeProductImage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: imageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: PRODUCT_NOT_FOUND | IMAGE_NOT_FOUND

  # ── Product (Admin) ──────────────────────────────────────────────────────────
  /api/admin/products/{id}/block:
    post:
      tags: [Product (Admin)]
      summary: Block product vi phạm chính sách
      description: Emit `ProductBlockedEvent`. Seller không thể tự publish lại.
      operationId: blockProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'
  /api/admin/products/{id}/unblock:
    post:
      tags: [Product (Admin)]
      summary: Unblock product — chuyển về UNPUBLISHED
      description: Emit `ProductUnblockedEvent`.
      operationId: unblockProduct
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseVoid' }
        "404":
          $ref: '#/components/responses/404'

  # ── Search (Admin) ───────────────────────────────────────────────────────────
...